pyproject.toml
[project]
name = "scidatabot"
version = "0.1.0"
description = "scidatabot - 科学数据智能助手"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "loguru>=0.7.0",
    "typer>=0.12.0",
    "textual>=8.0.0",
    # LLM Providers
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    # HTTP Client
    "aiohttp>=3.9.0",
    # File Operations
    "aiofiles>=23.0.0",
    # Database
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    # Data Formats
    "openpyxl>=3.1.0",
    "pyarrow>=15.0.0",
    "netCDF4>=1.6.0",
    "h5py>=3.10.0",
    # Config
    "pyyaml>=6.0.0",
    # Monitoring
    "psutil>=5.9.0",
    # Cron
    "croniter>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
all = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "openpyxl>=3.1.0",
    "pyarrow>=15.0.0",
    "netCDF4>=1.6.0",
    "h5py>=3.10.0",
    "pyyaml>=6.0.0",
    "psutil>=5.9.0",
    "croniter>=2.0.0",
]

[project.scripts]
scidatabot = "scidatabot.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scidatabot"]

[tool.hatch.build.targets.sdist]
include = ["scidatabot/", "README.md"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
